Include empty result set error
This is an option string provided at the time of requesting a query.
Either by direct use of query() or via get_object() or get_objects()

If the string is not empty and no result is returned, then the internal
error() function will be called.

The default error function, which should always be replaced in an
extension, will output the text in its report.